Update assertion for newer annotation codegen
Previously, when `java_multiple_files` was set to `false`, and hence all message types came under an outer class, we annotated message types, not the outer class.

It did not make much sense, and was inconsistent with a similar case for annotating a Java package.

New codegen annotates an outer class.